+++ utils/storage-utils.js
@@ -3,7 +3,7 @@
  * 封装微信小程序的 Storage API，提供统一的存取接口和错误处理
  *
  * @module utils/storage-utils
- * @version 1.12.0
+ * @version 1.15.0
  * @license MIT
  * @author WuSuBuDuoMing
  */